Prevent Extra Refreshing of Untagged Shows
There is currently an issue with the TV Scanner that causes untagged shows to be scanned along with a scan targeted to a particular directory. The easy way to fix this is to move the refresh outside of the scan loop and do a generic refresh of the entire library.